Updated basic story for emptystate
