+++ apps/storybook/package.json
@@ -30,6 +30,8 @@
     "storybook": "^9.0.12",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.30.1",
-    "vite": "^6.3.5"
+    "vite": "^6.3.5",
+    "@mui/icons-material": "^7.0.0",
+    "@mui/material": "^7.0.0"
   }
 }
